+++ Changelog.md
@@ -7,6 +7,9 @@
 
 ### Ignition Transport 5.X.X
 
+1. Added the ability to Pause/Resume playback from the log replayer.
+   * [Pull request 334](https://bitbucket.org/ignitionrobotics/ign-transport/pull-requests/334)
+
 1. Added support for topic remapping when running "ign log playback". Note that
    the string ":=" is not allowed now as part of a partition, namespace or topic
    anymore.